trending-smc/requirements.txt
grpcio
grpcio-tools
ccxt
pandas
numpy